# specification/swagger.yaml
swagger: "2.0"
info:
  description: "FiatConnect provides a standardized server-side API specification for CICO providers to provide cash-in/cash-out functionality to clients."
  version: "0.1.0"
  title: "FiatConnect Specification"
...
schemes:
- "https"
- "http"
definitions:
...
  LoginErrorResponse:
    type: "object"
    properties:
      error:
        $ref: "#/definitions/LoginErrorEnum"
...
  LoginErrorEnum:
    type: "string"
    enum:
      - InvalidSignature
      - InvalidParameters
      - ContractLoginNotSupported
      - NonceInUse
      - IssuedTooEarly
      - ExpirationTooLong
...
paths:
...
  /auth/login:
    post:
      summary: "Log in with a SIWE message and get a session cookie"
      tags:
        - "auth"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "siwe"
          description: "SIWE object"
          schema:
            type: "object"
            required:
              - "message"
              - "signature"
            properties:
              message:
                type: "string"
                description: "The plaintext SIWE message"
              signature:
                type: "string"
                description: "An EIP-191 (for EOA) or EIP-1271 (for contract-owned accounts) signature for the SIWE message"
      responses:
        "200":
          description: "successful operation"
        "401":
          description: "failed operation"
          schema:
            $ref: "#/definitions/LoginErrorResponse"
...
securityDefinitions:
  siwe_auth:
    type: "basic"